refactor: rename hierarchy_introspection.py to field_path_detection.py
Rename file to better reflect its actual purpose after removing the
hierarchy system. The file contains FieldPathDetector which is about
automatic field path detection and type introspection, not hierarchies.

Changes:
- Rename openhcs/core/hierarchy_introspection.py → field_path_detection.py
- Update imports in pipeline_config.py and parameter_form_abstraction.py
- Update module and class docstrings to reflect actual purpose

The FieldPathDetector utility eliminates hardcoded field paths and
provides automatic type introspection for the UI system.
